+++ app/src/main/kotlin/org/dbtools/sample/roomsqlite/App.kt
@@ -1,6 +1,7 @@
 package org.dbtools.sample.roomsqlite
 
 import android.support.multidex.MultiDexApplication
+import com.jakewharton.threetenabp.AndroidThreeTen
 import timber.log.Timber
 
 
@@ -8,6 +9,7 @@ class App : MultiDexApplication() {
 
     override fun onCreate() {
         super.onCreate()
+        AndroidThreeTen.init(this)
         setupLogging()
     }