+++ app/src/main/kotlin/org/dbtools/sample/roomsqlite/datasource/database/main/individualview/IndividualView.kt
@@ -1,8 +1,0 @@
-
-package org.dbtools.sample.roomsqlite.datasource.database.main.individualview
-
-
-class IndividualView  {
-    var id: Long? = null
-    var name = ""
-}