+++ app/src/main/kotlin/org/dbtools/sample/roomsqlite/datasource/database/main/type/IndividualType.kt
@@ -1,5 +1,0 @@
-package org.dbtools.sample.roomsqlite.datasource.database.main.type
-
-enum class IndividualType {
-    HEAD, SPOUSE, CHILD
-}