+++ app/src/main/kotlin/org/dbtools/sample/roomsqlite/model/db/main/individualview/IndividualView.kt
@@ -1,0 +1,8 @@
+
+package org.dbtools.sample.roomsqlite.model.db.main.individualview
+
+
+class IndividualView  {
+    var id: Long? = null
+    var name = ""
+}