+++ app/src/main/kotlin/org/dbtools/sample/roomsqlite/model/db/main/type/IndividualType.kt
@@ -1,0 +1,5 @@
+package org.dbtools.sample.roomsqlite.model.db.main.type
+
+enum class IndividualType {
+    HEAD, SPOUSE, CHILD
+}